Update classic infra API Key
In the console, if you select your classic infra API key, there is not an option for `username`, only description, email address, date created, created by, and the API Key Value. In my case, my classic infra username is the same as my "VPN" username shown in the IAM overview for my user, but I doubt this is standard.